feat(ttl): 实现过期命令与主动清理机制

+++ src/main/java/cn/twopair/resp/RespInt.java
@@ -12,5 +12,6 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class RespInt implements Resp {
-	private final int value;
+	// RESP 整数是有符号 64 位整数，使用 long。
+	private final long value;
 }
